[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17724 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,14 +1,11 @@
 import java.util.Scanner;
 
-
-
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int a = scanner.nextInt();
-        int b = scanner.nextInt();
-        int answer = (a+b)*(a-b);
+        long a = scanner.nextLong();
+        long b = scanner.nextLong();
+        long answer = (a + b) * (a - b);
         System.out.println(answer);
     }
 }
-
